return CausalLMOutput
