+++ mttl/evaluators/loglike_evaluator.py
@@ -4,33 +4,7 @@ import tqdm
 
 from mttl.evaluators.base import Evaluator, switch_to_eval_mode
 from mttl.logging import logger
-
-
-def compute_loglike_loss(logits, labels, reduction="none"):
-    # calculate loss, could also be done inside of the model
-    bs = logits.size(0)
-    vocab_size = logits.size(-1)
-    labels = labels.squeeze(-1)
-    shift_logits = logits[..., :-1, :].contiguous()
-    shift_labels = labels[..., 1:].contiguous()
-
-    # Flatten the tokens
-    loss_fct = torch.nn.CrossEntropyLoss(reduction=reduction)
-    shift_logits = shift_logits.view(-1, vocab_size)
-    shift_labels = shift_labels.view(-1)
-
-    # Enable model parallelism
-    shift_labels = shift_labels.to(shift_logits.device)
-    loss = loss_fct(shift_logits, shift_labels)
-
-    # reshape back
-    if reduction == "none":
-        loss = loss.view((bs, -1))
-        # mean only non-zero
-        non_zero_loss = (loss != 0).sum(dim=-1)
-        non_zero_loss[non_zero_loss == 0] = 1
-        loss = loss.sum(dim=-1) / non_zero_loss
-    return loss
+from mttl.models.utils import compute_loglike_loss
 
 
 class LogLikeEvaluator(Evaluator):
@@ -82,17 +56,16 @@ class LogLikeEvaluator(Evaluator):
 
             with torch.no_grad():
                 if isinstance(model, ExpertModule) or isinstance(model, ExpertModel):
-                    loss_per_option, _ = model.forward(**batch, reduction="none")
+                    logits = model.forward(**batch).logits
                 else:
                     logits = model.forward(
                         input_ids=batch["input_ids"],
                         attention_mask=batch["attention_mask"],
                     ).logits
 
-                    loss_per_option = compute_loglike_loss(
-                        logits, batch["labels"], reduction="none"
-                    )
-
+                loss_per_option = compute_loglike_loss(
+                    logits, batch["labels"], reduction="none"
+                )
                 loss_per_option = loss_per_option.cpu().numpy()
                 loss_per_example = [
                     loss_per_option[
