+++ mttl/models/base_model.py
@@ -14,6 +14,7 @@ import torch
 from attrs import field
 from huggingface_hub import hf_hub_download
 from torch.optim.optimizer import Optimizer
+from transformers.modeling_outputs import CausalLMOutput
 
 from mttl.arguments import Args, ExpertConfig, MoEExpertConfig, MultiExpertConfig
 from mttl.logging import logger
@@ -175,37 +176,11 @@ class BaseExpertModel(torch.nn.Module, Registrable):
         input_ids,
         attention_mask=None,
         labels=None,
-        reduction=None,
         **kwargs,
-    ):
-        outputs = self.model.forward(input_ids, attention_mask=attention_mask, **kwargs)
-
-        if labels is not None:
-            # calculate loss, could also be done inside of the model
-            bs = input_ids.size(0)
-            logits = outputs.logits
-            vocab_size = logits.size(-1)
-            labels = labels.squeeze(-1)
-            shift_logits = logits[..., :-1, :].contiguous()
-            shift_labels = labels[..., 1:].contiguous()
-
-            # Flatten the tokens
-            loss_fct = torch.nn.CrossEntropyLoss(reduction=reduction)
-            shift_logits = shift_logits.view(-1, vocab_size)
-            shift_labels = shift_labels.view(-1)
-
-            # Enable model parallelism
-            shift_labels = shift_labels.to(shift_logits.device)
-            loss = loss_fct(shift_logits, shift_labels)
-
-            # reshape back
-            if reduction == "none":
-                loss = loss.view((bs, -1))
-                # mean only non-zero
-                non_zero_loss = (loss != 0).sum(dim=-1)
-                non_zero_loss[non_zero_loss == 0] = 1
-                loss = loss.sum(dim=-1) / non_zero_loss
-            return loss, outputs
+    ) -> CausalLMOutput:
+        outputs = self.model.forward(
+            input_ids, attention_mask=attention_mask, labels=labels, **kwargs
+        )
         return outputs
 
     @property
